Move ProjectConfig and ProjectInfo to ProjectAttachReply.kt and rename it ProjectRelatedClasses.kt for clarity.
