+++ android/BOINC/app/src/main/java/edu/berkeley/boinc/rpc/ProjectRelatedClasses.kt
@@ -20,6 +20,9 @@ package edu.berkeley.boinc.rpc
 
 import android.os.Parcel
 import android.os.Parcelable
+import java.io.Serializable
+
+data class ProjectAttachReply(var errorNum: Int = 0, val messages: MutableList<String> = mutableListOf())
 
 // according to http://boinc.berkeley.edu/trac/wiki/WebRpc
 data class ProjectConfig(
@@ -42,8 +45,8 @@ data class ProjectConfig(
 ) : Parcelable {
     private constructor(parcel: Parcel) :
             this(parcel.readInt(), parcel.readString() ?: "", parcel.readString() ?: "",
-                    parcel.readString() ?: "", parcel.readString() ?: "", parcel.readInt(),
-                    parcel.readInt(), parcel.readString() ?: "") {
+                    parcel.readString() ?: "", parcel.readString() ?: "",
+                    parcel.readInt(), parcel.readInt(), parcel.readString() ?: "") {
         parcel.readList(platforms, PlatformInfo::class.java.classLoader)
         termsOfUse = parcel.readString()
         val bArray = parcel.createBooleanArray()!!
@@ -102,6 +105,59 @@ data class ProjectConfig(
             override fun createFromParcel(parcel: Parcel) = ProjectConfig(parcel)
 
             override fun newArray(size: Int) = arrayOfNulls<ProjectConfig>(size)
+        }
+    }
+}
+
+// needs to be serializable to be put into Activity start Intent
+data class ProjectInfo(
+        var name: String = "",
+        var url: String = "",
+        var generalArea: String? = null,
+        var specificArea: String? = null,
+        var description: String? = null,
+        var home: String? = null,
+        var platforms: List<String> = mutableListOf(),
+        var imageUrl: String? = null,
+        var summary: String? = null
+) : Serializable, Parcelable {
+    @Suppress("UNCHECKED_CAST")
+    private constructor(parcel: Parcel) :
+            this(parcel.readString() ?: "", parcel.readString() ?: "", parcel.readString(),
+                    parcel.readString(), parcel.readString(), parcel.readString(),
+                    parcel.readSerializable() as ArrayList<String>, parcel.readString(), parcel.readString())
+
+    override fun describeContents() = 0
+
+    override fun writeToParcel(dest: Parcel, arg1: Int) {
+        dest.writeString(name)
+        dest.writeString(url)
+        dest.writeString(generalArea)
+        dest.writeString(specificArea)
+        dest.writeString(description)
+        dest.writeString(home)
+        dest.writeSerializable(ArrayList(platforms))
+        dest.writeString(imageUrl)
+        dest.writeString(summary)
+    }
+
+    object Fields {
+        const val GENERAL_AREA = "general_area"
+        const val SPECIFIC_AREA = "specific_area"
+        const val HOME = "home"
+        const val PLATFORMS = "platforms"
+        const val IMAGE_URL = "image_url"
+        const val SUMMARY = "summary"
+    }
+
+    companion object {
+        private const val serialVersionUID = -5944047529950035455L // auto generated
+
+        @JvmField
+        val CREATOR: Parcelable.Creator<ProjectInfo> = object : Parcelable.Creator<ProjectInfo> {
+            override fun createFromParcel(parcel: Parcel) = ProjectInfo(parcel)
+
+            override fun newArray(size: Int) = arrayOfNulls<ProjectInfo>(size)
         }
     }
 }
